move arg definitions out of gemma.h to app.h
