+++ util/app.h
@@ -98,6 +98,125 @@ class AppArgs : public ArgsBase<AppArgs> {
   }
 };
 
+struct LoaderArgs : public ArgsBase<LoaderArgs> {
+  LoaderArgs(int argc, char* argv[]) { InitAndParse(argc, argv); }
+
+  static std::string ToLower(const std::string& text) {
+    std::string result = text;
+    std::transform(begin(result), end(result), begin(result),
+                   [](unsigned char c) { return std::tolower(c); });
+    return result;
+  }
+
+  gcpp::Model ModelType() const {
+    const std::string model_type_lc = ToLower(model_type);
+    if (model_type_lc == "2b-pt" || model_type_lc == "2b-it") {
+      return gcpp::Model::GEMMA_2B;
+    } else {
+      return gcpp::Model::GEMMA_7B;
+    }
+  }
+
+  gcpp::ModelTraining ModelTraining() const {
+    const std::string model_type_lc = ToLower(model_type);
+    if (model_type_lc == "7b-pt" || model_type_lc == "2b-pt") {
+      return gcpp::ModelTraining::GEMMA_PT;
+    } else {
+      return gcpp::ModelTraining::GEMMA_IT;
+    }
+  }
+
+  // Returns error string or nullptr if OK.
+  const char* Validate() const {
+    const std::string model_type_lc = ToLower(model_type);
+    if (model_type.empty()) {
+      return "Missing --model flag, need to specify either 2b-pt, 7b-pt, "
+             "2b-it, or 7b-it.";
+    }
+    if (model_type_lc != "2b-pt" && model_type_lc != "7b-pt" &&
+        model_type_lc != "2b-it" && model_type_lc != "7b-it") {
+      return "Model type must be 2b-pt, 7b-pt, 2b-it, or "
+             "7b-it.";
+    }
+    if (tokenizer.path.empty()) {
+      return "Missing --tokenizer flag, a file for the tokenizer is required.";
+    }
+    if (compressed_weights.path.empty()) {
+      return "Missing --compressed_weights flag, a file for the compressed "
+             "model.";
+    }
+    return nullptr;
+  }
+
+  Path tokenizer;
+  Path weights;             // uncompressed weights file location
+  Path compressed_weights;  // compressed weights file location
+  std::string model_type;
+
+  template <class Visitor>
+  void ForEach(const Visitor& visitor) {
+    visitor(tokenizer, "tokenizer", Path(),
+            "Path name of tokenizer model file.\n    Required argument.");
+    visitor(
+        compressed_weights, "compressed_weights", Path(),
+        "Path name of compressed weights file, regenerated from `--weights` "
+        "file if "
+        "the compressed weights file does not exist.\n    Required argument.");
+    visitor(model_type, "model", std::string(),
+            "Model type\n    2b-it = 2B parameters, instruction-tuned\n    "
+            "2b-pt = 2B parameters, pretrained\n    7b-it = 7B parameters "
+            "instruction-tuned\n    7b-pt = 7B parameters, pretrained\n"
+            "    Required argument.");
+    visitor(weights, "weights", Path(),
+            "Path name of model weights (.sbs) file. Only required if "
+            "compressed_weights file is not present and needs to be "
+            "regenerated. This parameter is only required for compressing"
+            "new model weight exports, otherwise it is not needed.");
+  }
+};
+
+
+struct InferenceArgs : public ArgsBase<InferenceArgs> {
+  InferenceArgs(int argc, char* argv[]) { InitAndParse(argc, argv); }
+
+  size_t max_tokens;
+  size_t max_generated_tokens;
+
+  float temperature;
+  bool deterministic;
+  bool multiturn;
+
+  // Returns error string or nullptr if OK.
+  const char* Validate() const {
+    if (max_tokens > gcpp::kSeqLen) {
+      return "max_tokens is larger than the maximum sequence length (see "
+             "configs.h).";
+    }
+    if (max_generated_tokens > max_tokens) {
+      return "Maximum number of generated tokens is larger than the maximum "
+             "total tokens.";
+    }
+    return nullptr;
+  }
+
+  template <class Visitor>
+  void ForEach(const Visitor& visitor) {
+    visitor(max_tokens, "max_tokens", size_t{3072},
+            "Maximum number of tokens in prompt + generation.");
+    visitor(max_generated_tokens, "max_generated_tokens", size_t{2048},
+            "Maximum number of tokens to generate.");
+
+    visitor(temperature, "temperature", 1.0f, "Temperature for top-K", 2);
+    visitor(deterministic, "deterministic", false,
+            "Make top-k sampling deterministic", 2);
+    visitor(multiturn, "multiturn", false,
+            "Multiturn mode\n    0 = clear KV cache after every "
+            "interaction\n    1 = continue KV cache after every interaction\n  "
+            "  Default : 0 (conversation "
+            "resets every turn)");
+  }
+};
+
 }  // namespace gcpp
 
 #endif  // THIRD_PARTY_GEMMA_CPP_UTIL_APP_H_
